base.py - path_like -> PathLike; move numpy dependent functions into numpy_util.py
